pyproject.toml
[tool.poetry]
name = "ltnexp03"
version = "0.1.0"
description = "A minimal Python microservice"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
z3-solver = "^4.15.1.0"
ltntorch = "^1.0.2"
scikit-learn = "^1.7.0"
torch = "^2.7.1"
matplotlib = "^3.10.3"
wn = "^0.13.0"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.9.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "app.main:start_server"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3
